pyproject.toml
[tool.poetry]
name = "simtrans"
version = "0.1.0"
description = ""
authors = ["meloidae <meloidaee@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.7,<3.10"
sacrebleu = {extras = ["ja"], version = "^1.4.14"}
hydra-core = "^1.0.5"
sentencepiece = "^0.1.95"
cloudpickle = "^1.6.0"
torch = "^1.7.1"
transformers = "^4.2.2"
boto3 = "^1.17.12"
kytea = "^0.1.5"
numba = "^0.53.0"
nltk = "^3.6.1"
jupyter = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
